projects/etl-pipeline-covid-api/visualizations/pyproject.toml
[tool.poetry]
name = "covid-dashboard"
version = "0.1.0"
description = "Streamlit dashboard for COVID-19 stats"
authors = ["codesVarun <singhvarun9554@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
streamlit = "^1.35.0"
plotly = "^5.22.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"